Update global_sensitivity.md
Documentation on the global sensitivity analysis is currently quite sparse.

The analysis does not work with models `f` that return a single output. The information that the model needs to output a vector (instead of a single value) would have saved me time. Therefore, I suggest to add it here.