introduction to adjoint sensitivity analysis
I have tried to make the use of formulae more consistent. In particular, the arguments of the functions should be clear, using u(p), u(t), u(t,p) interchangeably is too confusing. There is still more work to do in this direction, however, I would like to see whether you agree with the PR. The term "functional" should not be used for ``g(u,t,p)`` but for ``G(u,p)`` (not yet edited).