Update spiking_neural_systems.md
It looks odd, when the membrane potential in an integrate-and-fire neuron overshoots the threshold.

Btw: I find it convenient to use `ComponentArrays` for parameters, such that I can write e.g. `integrator.p.Vth` instead of `integrator.p[4]`, but I didn't apply this change here, as it would introduce a new dependency.