Minor rewording of explanation of "diagonal vs non-diagonal noise" in SDE examples.
I tried to polish the explanation of the diagonal vs non-diagonal noise in SDE examples. In particular, I found it confusing that the function `g` in the code for the diagonal noise returned/modified a vector, while the equally named g in the equation was a matrix. 

While working on the rewording, I also replaced in the codes the names of in-place functions `f` and `g` with `f!` and `g!`. I hope this will not be found as annoying nitpicking.